[refactor] user: 서비스 Command, Query로 분리
