+++ src/main/java/org/example/ootoutfitoftoday/domain/user/service/query/UserQueryServiceImpl.java
@@ -1,7 +1,6 @@
-package org.example.ootoutfitoftoday.domain.user.service;
+package org.example.ootoutfitoftoday.domain.user.service.query;
 
 import lombok.RequiredArgsConstructor;
-import org.example.ootoutfitoftoday.domain.user.entity.User;
 import org.example.ootoutfitoftoday.domain.user.repository.UserRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -10,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class UserServiceImpl implements UserService {
+public class UserQueryServiceImpl implements UserQueryService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
@@ -37,35 +36,5 @@ public class UserServiceImpl implements UserService {
     @Transactional(readOnly = true)
     public boolean existsByPhoneNumber(String phoneNumber) {
         return userRepository.existsByPhoneNumber(phoneNumber);
-    }
-
-    @Override
-    public void save(User user) {
-        userRepository.save(user);
-    }
-
-    @Override
-    public void initializeAdmin(
-            String loginId,
-            String email,
-            String nickname,
-            String username,
-            String password,
-            String phoneNumber
-    ) {
-        if (!existsByLoginId(loginId)) {
-            User admin = User.createAdmin(
-                    loginId,
-                    email,
-                    nickname,
-                    username,
-                    passwordEncoder.encode(password),
-                    phoneNumber
-            );
-            save(admin);
-            System.out.println("관리자 계정 초기 생성 완료되었습니다.");
-        } else {
-            System.out.println("관리자 계정이 이미 존재합니다.");
-        }
     }
 }
